app/requirements.txt
asgiref==3.11.0
Django>=4.2
djangorestframework==3.16.1
gunicorn==24.1.1
iniconfig==2.3.0
packaging==26.0
pluggy==1.6.0
psycopg2-binary==2.9.11
Pygments==2.19.2
pytest==9.0.2
pytest-django==4.11.1
python-decouple==3.8
sqlparse==0.5.5
pytest>=7.2.0
pytest-django>=4.5.0
